requirements.txt
requests>=2.31.0
openpyxl>=3.1.0
sentence-transformers>=2.2.2
torch>=2.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
